con frontEnd

+++ CondominioAPI/Data/Entities/CobroEntity.cs
@@ -14,7 +14,7 @@ namespace CondominioAPI.Data.Entities
         public long Id { get; set; }
         public string Descripcion { get; set; }
         public float? Valor { get; set; }
-        public double? MultaPorcentaje { get; set; }
+        public float? MultaPorcentaje { get; set; }
         public DateTime? FechaRegistro { get; set; }
         public DateTime? FechaVencimiento { get; set; }
         [ForeignKey("DepartamentoId")]
